create degree - department

+++ src/main/java/com/example/gradle/entity/Degree.java
@@ -22,5 +22,5 @@ public class Degree extends BaseString {
     @Column(name="name")
     private String name;
     @Column(name="value")
-    private Integer value;
+    private Float value;
 }
